Extract native database browsing and learned skills endpoints
Move request-scoped database and learned-skill callbacks into cohesive private FFI modules, preserving store brackets and terminal callback behavior.
