+++ packages/agent-native-bridge/agent-native-bridge.cabal
@@ -79,6 +79,8 @@ foreign-library haskell-agent-bridge
                     , Agent.CLI.MacOS.NativeGatewayBoundary
                     , Agent.CLI.MacOS.NativeModelCatalog
                     , Agent.CLI.MacOS.McpAdminBridge
+                    , Agent.CLI.MacOS.DatabaseBrowseBridge
+                    , Agent.CLI.MacOS.LearnedSkillsBridge
                     , Agent.CLI.MacOS.RepositoryInput
                     , Agent.CLI.MacOS.ResourceAdmin
     c-sources:        cbits/HaskellAgentBridgeRuntime.c
@@ -167,6 +169,8 @@ test-suite agent-native-bridge-test
                      , Agent.CLI.MacOS.NativeGatewayBoundary
                      , Agent.CLI.MacOS.NativeModelCatalog
                      , Agent.CLI.MacOS.McpAdminBridge
+                     , Agent.CLI.MacOS.DatabaseBrowseBridge
+                     , Agent.CLI.MacOS.LearnedSkillsBridge
                     , Agent.CLI.MacOS.ResourceAdmin
                      , Agent.CLI.MacOS.BrowserBridgeFFISpec
                      , Agent.CLI.MacOS.ComputerBridgeSpec
